requirements.txt
click>=8.0.0
pyyaml>=6.0
openai>=1.0.0
requests>=2.0.0
pytest>=7.0.0
